+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # CHANGELOG
 All notable changes to this project will be documented in this file.
 
+## [0.3.0] - 2019-09-11
+### Added
+- Database module
+- Database ID in configuration.ini
+- Name, age, nurse, and medication on front end
+
 ## [0.2.1] - 2019-09-11
 ### Removed
 - Today's min and max temperature